Moved tests to correct folder. Re-wrote broken tests with correct package pointers. Update and cleanup of pom.xml with correct dependencies and structure.
