+++ src/test/java/testrunner/InputNameTest.java
@@ -1,12 +1,12 @@
-package common;
+package testrunner;
 
-import junit.framework.TestCase;
+import org.junit.Assert;
 import org.junit.Test;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 
-public class InputNameTest extends TestCase {
+public class InputNameTest {
 
     @Test
     public void testInputName() {
@@ -14,12 +14,12 @@ public class InputNameTest extends TestCase {
         InputStream inputStream = new ByteArrayInputStream(inputName.getBytes());
         System.setIn(inputStream);
 
-        InputName inputNameInstance = new InputName();
+        common.InputName inputNameInstance = new common.InputName();
 
         String result = inputNameInstance.addCompetitor();
 
         System.setIn(System.in);
 
-        assertEquals("Umar", result);
+        Assert.assertEquals("Umar", result);
     }
 }
